requirements.txt
flask==1.1.2
numpy==1.19.2
pandas==1.1.3
Pillow==8.0.1
requests==2.24.0
torch==1.7.1
torchsummary==1.5.1
torchvision==0.8.2
ipython==7.19.0